requirements.txt
dspy-ai>=2.4.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
scikit-learn>=1.3.0
gradio>=4.44.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
scipy>=1.11.0
statsmodels>=0.14.0
jupyter>=1.0.0
ipykernel>=6.25.0
python-dateutil>=2.8.0
regex>=2023.8.8
typing-extensions>=4.7.0
pydantic>=2.0.0
kaleido>=0.2.1
reportlab>=4.0.0
weasyprint>=60.0
jinja2>=3.1.0
pathlib>=1.0.0
google-cloud-aiplatform==1.71.1